refactor(analysis): move src unit tests to integration tests
- move src/unit_tests to foundation/analysis/tests with flat filenames

- externalize geometry/numerics tests from src

- remove empty legacy unit_tests directories
