+++ foundation/analysis/src/numerics/integration.rs
@@ -120,57 +120,3 @@ where
 
     sum * h
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[derive(Clone)]
-    struct TestVector2D {
-        x: f64,
-        y: f64,
-    }
-
-    impl NormedVector for TestVector2D {
-        fn norm(&self) -> f64 {
-            (self.x * self.x + self.y * self.y).sqrt()
-        }
-    }
-
-    #[test]
-    fn test_newton_arc_length_line() {
-        // 直線の弧長：r'(t) = (1, 0) でt∈[0,1]
-        let line_derivative = |_t: f64| TestVector2D { x: 1.0, y: 0.0 };
-        let arc_length = newton_arc_length(line_derivative, 0.0, 1.0, 100);
-
-        // 理論値: 1.0
-        assert!((arc_length - 1.0).abs() < 1e-10);
-    }
-
-    #[test]
-    fn test_newton_arc_length_circle() {
-        // 半径1の円の弧長：r'(t) = (-sin(t), cos(t))
-        let circle_derivative = |t: f64| TestVector2D {
-            x: -t.sin(),
-            y: t.cos(),
-        };
-
-        // 半周の弧長を計算（理論値: π）
-        let arc_length = newton_arc_length(circle_derivative, 0.0, std::f64::consts::PI, 1000);
-        assert!((arc_length - std::f64::consts::PI).abs() < 1e-3);
-    }
-
-    #[test]
-    fn test_trapezoidal_rule_quadratic() {
-        // ∫₀¹ x² dx = 1/3 を計算
-        let result = trapezoidal_rule(|x| x * x, 0.0, 1.0, 1000);
-        assert!((result - 1.0 / 3.0).abs() < 1e-6);
-    }
-
-    #[test]
-    fn test_trapezoidal_rule_sine() {
-        // ∫₀π sin(x) dx = 2 を計算
-        let result = trapezoidal_rule(|x| x.sin(), 0.0, std::f64::consts::PI, 10000);
-        assert!((result - 2.0).abs() < 1e-4);
-    }
-}
