--- foundation/analysis/src/unit_tests/mod.rs
@@ -1,14 +1,0 @@
-//! analysis クレートのユニットテスト
-//!
-//! このモジュールは、analysisクレートの各機能のテストコードを含む。
-//! 元々は各ソースファイル内にあった #[cfg(test)] ブロックを
-//! 独立したテストファイルに分離している。
-pub mod interpolation_additional_tests;
-pub mod interpolation_tests;
-// pub mod linalg; - linalg配下のテストと重複するため無効化
-pub mod numerical_additional_tests;
-pub mod numerical_tests;
-// pub mod sampling_additional_tests; - 機能はgeo_algorithmsに移動済み
-// pub mod sampling_tests; - 機能はgeo_algorithmsに移動済み
-pub mod scalar_tests;
-pub mod statistics_tests;
